pyproject.toml
[project]
name = "gasprices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arrow>=1.3.0",
    "debugpy>=1.8.16",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "plotly[express]>=6.3.0",
    "polars>=1.32.3",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "selectolax>=0.3.33",
    "uvicorn>=0.35.0",
]
